move transaction to types
