+++ packages/txservice/src/dispatch.ts
@@ -2,11 +2,10 @@ import { BigNumber, Signer, providers, utils } from "ethers";
 import PriorityQueue from "p-queue";
 import { createLoggingContext, delay, getUuid, jsonifyError, Logger, RequestContext } from "@connext/nxtp-utils";
 
-import { Gas, WriteTransaction } from "./types";
+import { Gas, WriteTransaction, Transaction } from "./types";
 import { BadNonce, TransactionReplaced, TransactionReverted, TimeoutError, TransactionServiceFailure } from "./error";
 import { ChainConfig, TransactionServiceConfig } from "./config";
 import { ChainRpcProvider } from "./provider";
-import { Transaction } from "./transaction";
 
 type DispatchCallbacks = {
   onSubmit: (transaction: Transaction) => void;
